fix: adding code to disable aggregation
