+++ rudderreporter/src/main/java/com/rudderstack/android/ruddermetricsreporterandroid/Syncer.kt
@@ -26,4 +26,5 @@ interface Syncer {
     fun stopScheduling()
 
     fun flushAllMetrics()
+
 }
