+++ rudderreporter/src/main/java/com/rudderstack/android/ruddermetricsreporterandroid/internal/metrics/DefaultAggregatorHandler.kt
@@ -20,13 +20,17 @@ import com.rudderstack.android.ruddermetricsreporterandroid.metrics.LongCounter
 import com.rudderstack.android.ruddermetricsreporterandroid.metrics.LongGauge
 import com.rudderstack.android.ruddermetricsreporterandroid.metrics.MetricModel
 import com.rudderstack.android.ruddermetricsreporterandroid.metrics.MetricType
+import java.util.concurrent.atomic.AtomicBoolean
 
-class DefaultAggregatorHandler(private val reservoir: Reservoir) : AggregatorHandler {
+class DefaultAggregatorHandler(private val reservoir: Reservoir,
+isEnabled: Boolean = true) : AggregatorHandler {
+    private val _isEnabled = AtomicBoolean(isEnabled)
     override fun LongCounter.recordMetric(value: Long) {
         recordMetric(value, mapOf())
     }
 
     override fun LongCounter.recordMetric(value: Long, attributes: Map<String,String>) {
+        if(!_isEnabled.get()) return
         reservoir.insertOrIncrement(MetricModel(name, MetricType.COUNTER,
             value, attributes))
     }
@@ -36,7 +40,12 @@ class DefaultAggregatorHandler(private val reservoir: Reservoir) : AggregatorHan
     }
 
     override fun LongGauge.recordMetric(value: Long, attributes: Map<String,String>) {
+        if(!_isEnabled.get()) return
         reservoir.insertOrIncrement(MetricModel(name, MetricType.GAUGE,
             value, attributes))
+    }
+
+    override fun enable(enable: Boolean) {
+        _isEnabled.set(enable)
     }
 }
